+++ src/fields/AddressField.php
@@ -19,6 +19,7 @@ use craft\elements\db\ElementQuery;
 use craft\elements\db\ElementQueryInterface;
 use craft\elements\Entry;
 use craft\events\CancelableEvent;
+use craft\events\DefineFieldKeywordsEvent;
 use craft\events\PopulateElementEvent;
 use craft\helpers\Json;
 use doublesecretagency\googlemaps\enums\Defaults;
@@ -683,12 +684,25 @@ class AddressField extends Field implements PreviewableFieldInterface
      */
     public function getSearchKeywords(mixed $value, ElementInterface $element): string
     {
-        // If no value, bail with empty string
-        if (!$value) {
+        // If no value or not an Address
+        if (!$value || !is_a($value, AddressModel::class)) {
+            // Bail with empty string
             return '';
         }
 
-        /** @var AddressModel $value */
+        // Give plugins/modules a chance to define custom keywords
+        if ($this->hasEventHandlers(self::EVENT_DEFINE_KEYWORDS)) {
+            $event = new DefineFieldKeywordsEvent([
+                'value' => $value,
+                'element' => $element,
+            ]);
+            $this->trigger(self::EVENT_DEFINE_KEYWORDS, $event);
+            if ($event->handled) {
+                return $event->keywords;
+            }
+        }
+
+        // Return default keywords
         return implode(' ', [
             $value->formatted,
             $value->name,